Split Websocket into a plaintext and TLS implementation.
